removed throws InterruptedException from all methods under class PlayListTests
