+++ src/test/java/BaseTest.java
@@ -2,30 +2,18 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chromium.ChromiumNetworkConditions;
-import org.openqa.selenium.chromium.HasNetworkConditions;
-import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.interactions.Actions;
-import org.openqa.selenium.remote.Augmenter;
-import org.openqa.selenium.remote.DesiredCapabilities;
-import org.openqa.selenium.remote.RemoteWebDriver;
-import org.openqa.selenium.safari.SafariDriver;
-import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.annotations.*;
 
 import java.net.MalformedURLException;
-import java.net.URI;
 import java.time.Duration;
 
 public class BaseTest {
-
     WebDriver driver;
     String url;
     WebDriverWait wait;
     Actions actions;
-
-
 
     @BeforeSuite
     public static void chromeConfigs() {
@@ -35,8 +23,6 @@ public class BaseTest {
         } else {
             System.setProperty("webdriver.chrome.driver", "chromedriver");
         }
-
-
     }
 
     @BeforeMethod
@@ -44,56 +30,15 @@ public class BaseTest {
     @Parameters({"baseURL"})
     // Make baseURL parameter optional, if it is null, then set it to something)
     public void launchBrowser(@Optional String baseURL) throws MalformedURLException {
-        if (baseURL == null)
-            baseURL ="https://bbb.testpro.io";
-        //driver = new ChromeDriver();
+        if (baseURL == null) baseURL = "https://bbb.testpro.io";
         System.setProperty("webdriver.gecko.driver", "geckodriver");
-        //driver = new FirefoxDriver();
-        //driver = new SafariDriver();
-        driver = pickBrowser(System.getProperty("browser"));
+        //driver = pickBrowser(System.getProperty("browser"));
+        driver = new ChromeDriver();
         actions = new Actions(driver);
-        // Make webdriver load the pages REALLY slow
-//        WebDriver augmentedDriver = new Augmenter().augment(driver);
-//        ChromiumNetworkConditions networkConditions = new ChromiumNetworkConditions();
-//        networkConditions.setDownloadThroughput(100 * 1024);
-//        networkConditions.setUploadThroughput(500 * 1024);
-//        networkConditions.setLatency(Duration.ofMillis(5000));
-//        ((HasNetworkConditions) augmentedDriver).setNetworkConditions(networkConditions);
-        // (comment out above lines to remove throttling)
-
-        // Wait for an element to show up for max of X seconds
-        // implicitlyWait(Duration.ofSeconds(60) will wait for UP to 60 seconds
-        // if element comes up after 1 second, it will move on
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
-
-        wait = new WebDriverWait(driver,Duration.ofSeconds(10));
-        // thread.sleep(60000) -- will wait 60s always
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(15));
+        wait = new WebDriverWait(driver, Duration.ofSeconds(10));
         url = baseURL;
         driver.get(url);
-
-    }
-
-    private WebDriver pickBrowser(String browser) throws MalformedURLException {
-        DesiredCapabilities caps = new DesiredCapabilities();
-        String gridURL = "http://192.168.1.2:4444";
-        switch (browser){
-            case "firefox":
-                System.setProperty("webdriver.gecko.driver", "geckodriver");
-                return driver = new FirefoxDriver();
-            case "safari":
-                return driver = new SafariDriver();
-            case "grid-safari":
-                caps.setCapability("browserName", "safari");
-                return driver = new RemoteWebDriver(URI.create(gridURL).toURL(),caps);
-            case "grid-firefox":
-                caps.setCapability("browserName", "firefox");
-                return driver = new RemoteWebDriver(URI.create(gridURL).toURL(),caps);
-            case "grid-chrome":
-                caps.setCapability("browserName", "chrome");
-                return driver = new RemoteWebDriver(URI.create(gridURL).toURL(),caps);
-            default:
-                return driver = new ChromeDriver();
-        }
     }
 
     @AfterMethod
@@ -116,20 +61,14 @@ public class BaseTest {
         WebElement passwordField = driver.findElement(By.cssSelector("[type='password']"));
         passwordField.click();
         passwordField.sendKeys(password);
-
     }
 
-    @DataProvider(name="invalidCredentials")
-    public static Object[][] getCredentials(){
-
-        return new Object[][] {
-                {"invalid@class.com", "invalidPass"},
-                {"d@class.com", ""},
-                {"", ""}
-        };
+    @DataProvider(name = "invalidCredentials")
+    public static Object[][] getCredentials() {
+        return new Object[][]{{"invalid@class.com", "invalidPass"}, {"d@class.com", ""}, {"", ""}};
     }
 
-    public void login(){
+    public void login() {
         provideEmail("demo@class.com");
         providePassword("te$t$tudent");
         clickSubmitBtn();
