+++ src/test/java/Homework17.java
@@ -3,7 +3,7 @@ import org.openqa.selenium.WebElement;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-public class Homework17 extends BaseTest {
+public class Homework17 extends BaseTestOrigin {
