+++ src/test/java/Homework18.java
@@ -3,7 +3,7 @@ import org.openqa.selenium.WebElement;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-public class Homework18 extends BaseTest {
+public class Homework18 extends BaseTestOrigin {
 
     @Test
     public void playSong() throws InterruptedException {
