+++ src/test/java/Homework19.java
@@ -3,7 +3,7 @@ import org.openqa.selenium.WebElement;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-public class Homework19 extends BaseTest {
+public class Homework19 extends BaseTestOrigin {
 
     @Test
     public void deletePlaylist() throws InterruptedException {
